backend/requirements.txt
# Django & REST Framework
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.2.0
pytesseract==0.3.10
pdf2image==1.16.3

# HTTP Requests (for Ollama)
requests==2.31.0
httpx==0.26.0

# Utilities
python-decouple==3.8
python-dotenv==1.0.0

# Validation
pydantic==2.5.3

# Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0

# Code Quality
flake8==7.0.0
black==23.12.1
isort==5.13.2
